pyproject.toml
[project]
name = "google-next-mortgage-ai"
version = "0.1.0"
description = "Mortgage AI processing using Google Cloud and Gemini"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Web Framework & Server
    "flask==3.1.2",
    "flask-wtf==1.2.2",
    "gunicorn==23.0.0",
    "werkzeug==3.1.3",

    # Google Cloud & AI
    "google-cloud-documentai==3.6.0",
    "google-cloud-firestore==2.21.0",
    "google-cloud-vertexai>=0.0.1",
    "google-cloud-aiplatform>=1.38.0",
    "google-genai==1.44.0",
    "google-generativeai",

    # Document Processing
    "pypdf2==3.0.1",
    "pdf2image==1.17.0",
    "pillow==12.0.0",
    "reportlab==4.2.5",

    # Data & Environment
    "pydantic==2.12.2",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv==1.1.1",
]

# -- PyTorch CPU Configuration --
# This adds the CPU-only wheel index to save space (crucial for Replit/Nix)
[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
# Only map packages that explicitly need the CPU index here.
# Since your dependencies list doesn't currently include 'torch', 
# this section is empty. If you add torch later, uncomment the lines below:
# torch = [{ index = "pytorch-cpu" }]
# torchvision = [{ index = "pytorch-cpu" }]